# requirements.txt.txt
# requirements.txt
PyMuPDF==1.24.10
sentence-transformers==3.1.1
faiss-cpu==1.9.0
openai==1.47.1
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
numpy==2.1.1
sqlite3